+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to
 
 - Show collaborative editor toggle (beaker icon) when creating new workflows
   [#3797](https://github.com/OpenFn/lightning/pull/3797)
+- Mix task to merge project state files without database access
+  [#3615](https://github.com/OpenFn/lightning/issues/3615)
 
 ### Changed